refactored wall
